requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
python-multipart==0.0.6
pydantic>=2.8.0
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-nmap==0.7.1
shodan==1.30.1
censys==2.2.8
whois==0.9.27
sqlparse==0.4.4
jinja2==3.1.2
reportlab==4.0.7
redis==5.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23